Rewrite & optimize mainlogic for NorthboundDatabase events mapping
